# Merge chunks together on compaction