--- a/compact.go
+++ b/compact.go
@@ -64,12 +64,12 @@ type Compactor interface {
 
 // LeveledCompactor implements the Compactor interface.
 type LeveledCompactor struct {
-	dir         string
-	metrics     *compactorMetrics
-	logger      log.Logger
-	ranges      []int64
-	chunkPool   chunkenc.Pool
-	mergeChunks bool
+	dir              string
+	metrics          *compactorMetrics
+	logger           log.Logger
+	ranges           []int64
+	chunkPool        chunkenc.Pool
+	mergeSmallChunks bool
 }
 
 type compactorMetrics struct {
@@ -135,11 +135,11 @@ func NewLeveledCompactor(r prometheus.Registerer, l log.Logger, ranges []int64, 
 		pool = chunkenc.NewPool()
 	}
 	return &LeveledCompactor{
-		ranges:      ranges,
-		chunkPool:   pool,
-		logger:      l,
-		metrics:     newCompactorMetrics(r),
-		mergeChunks: true,
+		ranges:           ranges,
+		chunkPool:        pool,
+		logger:           l,
+		metrics:          newCompactorMetrics(r),
+		mergeSmallChunks: true,
 	}, nil
 }
 
@@ -628,8 +628,8 @@ func (c *LeveledCompactor) populateBlock(blocks []BlockReader, meta *BlockMeta, 
 			}
 		}
 
-		if c.mergeChunks {
-			chks = mergeChunks(chks)
+		if c.mergeSmallChunks {
+			chks = c.mergeChunks(chks)
 		}
 
 		if err := chunkw.WriteChunks(chks...); err != nil {
@@ -688,48 +688,49 @@ func (c *LeveledCompactor) populateBlock(blocks []BlockReader, meta *BlockMeta, 
 	return nil
 }
 
-func mergeChunks(chks []chunks.Meta) []chunks.Meta {
+func (c *LeveledCompactor) mergeChunks(chks []chunks.Meta) []chunks.Meta {
 	newChks := make([]chunks.Meta, 0, len(chks))
 	for i := 0; i < len(chks); i++ {
-		if i < len(chks)-1 && chks[i].Chunk.NumSamples()+chks[i+1].Chunk.NumSamples() <= 480 {
-			newChunk := chunkenc.NewXORChunk()
-			app, err := newChunk.Appender()
-			if err != nil {
-				return chks
-			}
-			it := chks[i].Chunk.Iterator()
-			for it.Next() {
-				app.Append(it.At())
-			}
-			if it.Err() != nil {
-				return chks
-			}
-			minTime := chks[i].MinTime
-			maxTime := chks[i].MaxTime
-			ref := chks[i].Ref
-
-			for i < len(chks)-1 && newChunk.NumSamples()+chks[i+1].Chunk.NumSamples() <= 480 {
-				it = chks[i+1].Chunk.Iterator()
-				for it.Next() {
-					app.Append(it.At())
-				}
-				if it.Err() != nil {
-					return chks
-				}
-				maxTime = chks[i+1].MaxTime
-				i++
-			}
-			newChks = append(newChks, chunks.Meta{
-				MinTime: minTime,
-				MaxTime: maxTime,
-				Ref:     ref,
-				Chunk:   newChunk,
-			})
-		} else {
+		if i >= len(chks)-1 || chks[i].Chunk.NumSamples()+chks[i+1].Chunk.NumSamples() > 480 {
 			newChks = append(newChks, chks[i])
+			continue
 		}
+		newChunk := chunkenc.NewXORChunk()
+		app, err := newChunk.Appender()
+		if err != nil {
+			level.Warn(c.logger).Log("msg", "failed to merge chunks", "err", err)
+			return chks
+		}
+		if err := appendChunk(app, chks[i].Chunk); err != nil {
+			level.Warn(c.logger).Log("msg", "failed to merge chunks", "err", err)
+			return chks
+		}
+		minTime := chks[i].MinTime
+		maxTime := chks[i].MaxTime
+
+		for i < len(chks)-1 && newChunk.NumSamples()+chks[i+1].Chunk.NumSamples() <= 480 {
+			if err := appendChunk(app, chks[i+1].Chunk); err != nil {
+				level.Warn(c.logger).Log("msg", "failed to merge chunks", "err", err)
+				return chks
+			}
+			maxTime = chks[i+1].MaxTime
+			i++
+		}
+		newChks = append(newChks, chunks.Meta{
+			MinTime: minTime,
+			MaxTime: maxTime,
+			Chunk:   newChunk,
+		})
 	}
 	return newChks
+}
+
+func appendChunk(app chunkenc.Appender, chunk chunkenc.Chunk) error {
+	it := chunk.Iterator()
+	for it.Next() {
+		app.Append(it.At())
+	}
+	return it.Err()
 }
 
 type compactionSeriesSet struct {