fixes: https://github.com/prometheus/tsdb/issues/236

This is a first pass at merging small chunks together into larger ones. Initial query benchmark results seem promising for when blocks become larger. This code currently takes the 120 samples per 2 hours plan used in the head and makes the chunk size proportionally larger as the block size increases.

benchcmp results between first and last commits:
```
benchmark                                                          old ns/op     new ns/op     delta
BenchmarkPersistedQueries/series=10,samplesPerSeries=1000-8        103098        106463        +3.26%
BenchmarkPersistedQueries/series=10,samplesPerSeries=10000-8       468642        259983        -44.52%
BenchmarkPersistedQueries/series=10,samplesPerSeries=100000-8      3987684       209855        -94.74%
BenchmarkPersistedQueries/series=100,samplesPerSeries=1000-8       863718        852374        -1.31%
BenchmarkPersistedQueries/series=100,samplesPerSeries=10000-8      4582881       2465040       -46.21%
BenchmarkPersistedQueries/series=100,samplesPerSeries=100000-8     47530956      1918334       -95.96%

benchmark                                                          old allocs     new allocs     delta
BenchmarkPersistedQueries/series=10,samplesPerSeries=1000-8        427            427            +0.00%
BenchmarkPersistedQueries/series=10,samplesPerSeries=10000-8       1977           1107           -44.01%
BenchmarkPersistedQueries/series=10,samplesPerSeries=100000-8      17267          877            -94.92%
BenchmarkPersistedQueries/series=100,samplesPerSeries=1000-8       3738           3738           +0.00%
BenchmarkPersistedQueries/series=100,samplesPerSeries=10000-8      19238          10538          -45.22%
BenchmarkPersistedQueries/series=100,samplesPerSeries=100000-8     172145         8238           -95.21%

benchmark                                                          old bytes     new bytes     delta
BenchmarkPersistedQueries/series=10,samplesPerSeries=1000-8        70499         70500         +0.00%
BenchmarkPersistedQueries/series=10,samplesPerSeries=10000-8       194376        123395        -36.52%
BenchmarkPersistedQueries/series=10,samplesPerSeries=100000-8      1317825       92106         -93.01%
BenchmarkPersistedQueries/series=100,samplesPerSeries=1000-8       341780        341780        +0.00%
BenchmarkPersistedQueries/series=100,samplesPerSeries=10000-8      1580558       870738        -44.91%
BenchmarkPersistedQueries/series=100,samplesPerSeries=100000-8     12815166      557849        -95.65%
```