@krasi-georgiev 2018-09-28
please remove the [WIP] when this is ready for a review.