@csmarchbanks 2018-10-13
@krasi-georgiev This is ready for a review now.

Sorry it took awhile, the last two weeks have been a bit crazy for me.