@krasi-georgiev 2018-10-15
yeah for now we need a PR against Prometheus, but let me review it first and will run the tests after. 
It will probably be a week or two before I get to this since we have few other PRs to finish before this one.