@csmarchbanks 2018-10-21
I added a unit test for `mergeChunks`, decided to leave the compaction tests in there to supplement the specific unit test.

I am going to think a bit more on how to simplify `mergeChunks`, I am not super happy with either code snippet yet...