@fabxc 2018-10-22
> makes the chunk size proportionally larger as the block size increases.

That sounds wrong. The compression ratio nearly reaches optimum average at about 60 samples. There's virtually no further gain at all beyond 120 samples ([Figure 6](http://www.vldb.org/pvldb/vol8/p1816-teller.pdf).
All we save is indexing those chunks but index size has not been a major concern IIRC.

Do we have any data that would back up that bigger chunks would become more efficient?
If you keep making chunks bigger and bigger, that would effectively mean spending a ton of CPU on each compaction to re-compress virtually every series, no?  
Queries will have to decompress day-long chunks just to access a small window of data within them. Thus a query for 1h and 4 weeks could take the same amount of time in unlucky cases.  

Please correct me if I misunderstood something.
This seems like an expensive and invasive operation and we'd need some hard data what improvement this shows. The added benchmark is relatively micro and doesn't reflect the big picture I think. There are some impressive improvements for sure – it would be good to verify whether those are owed to samples being actually compressed in one big sequence or whether this is a simple allocation problem that could just be optimized away.