@krasi-georgiev 2018-10-22
@csmarchbanks we can probably test all this roughly in a prombench test. If you open a PR against Prometheus with these changes I can start a test leave it running and test some query  timings, cpu usage etc.