@csmarchbanks 2018-10-22
@fabxc Your understanding is correct, definitely more cpu on compaction, and I don't know if it would make a real world difference. I wanted to get this in an ok enough place where it would make sense to test using prombench/a custom build of prometheus for a more realistic scenario, and verify if there are any noticeable improvements. If not, then I will close this PR because the complexity and extra cpu on compaction is definitely not worth it.

I will look into some allocation optimizations as well. That sounds like an interesting path to pursue.

@krasi-georgiev I will let you know when I make a prometheus PR. Should be sometime today.