@csmarchbanks 2018-10-23
Looking at the benchmarks in https://github.com/prometheus/prometheus/pull/4768, I don't think making bigger chunks is worthwhile. Query performance does not change much, and some compactions are taking much much longer (40 minutes vs 10 minutes).

An alternative could be to merge tiny chunks (say less than 40 samples) into another chunk to increase compression a bit. However, in my prometheus no more than 1% of chunks would be merged, so I doubt the extra complexity is worthwhile.